Updated creation of new projects to preserve user-provided name case.
Previously any project name with uppercase letters were converted to lowercase. This was to comply with requirements for NPM modules. However, we now updated the new project process to leave the name of the new directory for the project to match the case provided by the user, while still making the package name comply with NPM naming conventions. This means that if a user provides a name like so: composi -n "My Project", the package name would be "my-project", but the project folder would and html index title would remain "My Project".